163: Extract log entry verifier into separate class
Task-Url: https://github.com/exasol/exasol-testcontainers/issues/163
