+++ src/main/java/com/exasol/clusterlogs/LogPatternDetector.java
@@ -1,11 +1,6 @@
 package com.exasol.clusterlogs;
 
-import java.io.*;
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.util.TimeZone;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.io.IOException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -18,14 +13,11 @@ import com.exasol.containers.exec.ExitCode;
  */
 public class LogPatternDetector {
     private static final Logger LOGGER = LoggerFactory.getLogger(LogPatternDetector.class);
-    private static final Pattern LOG_ENTRY_PATTERN = Pattern
-            .compile("\\[. (\\d\\d)(\\d\\d)(\\d\\d) (\\d\\d:\\d\\d:\\d\\d).*");
     private final Container<? extends Container<?>> container;
     private final String logPath;
     private final String pattern;
     private final String logNamePattern;
-    private final TimeZone timeZone;
-    private final Instant afterUtc;
+    private final LogEntryPatternVerifier logEntryVerifier;
 
     /**
      * Create a new instance of the {@link LogPatternDetector}.
@@ -34,19 +26,16 @@ public class LogPatternDetector {
      * @param logPath        path of the log file to search
      * @param logNamePattern pattern used to find the file name
      * @param pattern        regular expression pattern for which to look out
-     * @param timeZone       time zone that serves as context for the short timestamps in the logs
-     * @param afterUtc       earliest time in the log after which the log message must appear
      */
     LogPatternDetector(final Container<? extends Container<?>> container, final String logPath,
-            final String logNamePattern, final String pattern, final TimeZone timeZone, final Instant afterUtc) {
+            final String logNamePattern, final String pattern, final LogEntryPatternVerifier logEntryVerifier) {
         this.container = container;
         this.logPath = logPath;
         this.logNamePattern = logNamePattern;
         this.pattern = pattern;
-        this.timeZone = timeZone;
-        this.afterUtc = afterUtc;
-        LOGGER.debug("Created log detector that scans for \"{}\" in \"{}/{}\" with time zone \"{}\"", pattern, logPath,
-                logNamePattern, timeZone.getDisplayName());
+        this.logEntryVerifier = logEntryVerifier;
+        LOGGER.debug("Created log detector that scans for \"{}\" in \"{}/{}\" with verifier {}", pattern, logPath,
+                logNamePattern, logEntryVerifier);
     }
 
     /**
@@ -86,38 +75,10 @@ public class LogPatternDetector {
                 "-name", this.logNamePattern, //
                 "-exec", "awk", "/" + this.pattern.replace("/", "\\/") + "/{a=$0}END{print a}", "{}", "+");
         if (result.getExitCode() == ExitCode.OK) {
-            return isLogMessageFound(result.getStdout());
+            return this.logEntryVerifier.isLogMessageFound(result.getStdout());
         } else {
             return false;
         }
-    }
-
-    private boolean isLogMessageFound(final String stdout) throws IOException {
-        final LocalDateTime afterLocal = convertUtcToLowResulionLocal(this.afterUtc);
-        try (final BufferedReader reader = new BufferedReader(new StringReader(stdout))) {
-            String line;
-            while ((line = reader.readLine()) != null) {
-                if (line.isBlank()) {
-                    continue;
-                }
-                final Matcher matcher = LOG_ENTRY_PATTERN.matcher(line);
-                if (matcher.matches()) {
-                    final String isoTimestamp = "20" + matcher.group(1) + "-" + matcher.group(2) + "-"
-                            + matcher.group(3) + "T" + matcher.group(4);
-                    final LocalDateTime timestamp = LocalDateTime.parse(isoTimestamp);
-                    if (timestamp.isAfter(afterLocal) || timestamp.isEqual(afterLocal)) {
-                        LOGGER.debug("Found matching log entry {} (after {}): {}", timestamp, afterLocal, line);
-                        return true;
-                    }
-                }
-            }
-            return false;
-        }
-    }
-
-    private LocalDateTime convertUtcToLowResulionLocal(final Instant afterUTC) {
-        final LocalDateTime localDateTime = LocalDateTime.ofInstant(afterUTC, this.timeZone.toZoneId());
-        return localDateTime.withNano(0);
     }
 
     /**
@@ -127,6 +88,6 @@ public class LogPatternDetector {
      */
     public String describe() {
         return "Scanning for log message pattern \"" + this.pattern + " in \"" + this.logPath + "/"
-                + this.logNamePattern + "\".";
+                + this.logNamePattern + "\". using " + this.logEntryVerifier;
     }
 }
